requirements.txt
attrdict==2.0.1
attrs==19.3.0
base58==2.0.0
cached-property==1.5.1
certifi==2019.11.28
chardet==3.0.4
cytoolz==0.10.1
Django==2.2.6
django-filter==2.2.0
django-model-utils==3.2.0
djangorestframework==3.10.3
eth-abi==2.1.0
eth-account==0.4.0
eth-bloom==1.0.3
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.1.2
eth-typing==2.2.1
eth-utils==1.8.4
hexbytes==0.2.0
idna==2.8
importlib-metadata==1.4.0
ipfshttpclient==0.4.12
jsonschema==3.2.0
lru-dict==1.1.6
Markdown==3.1.1
more-itertools==8.1.0
multiaddr==0.0.9
mypy-extensions==0.4.3
netaddr==0.7.19
parsimonious==0.8.1
protobuf==3.11.2
py-ecc==1.7.1
pycryptodome==3.9.4
pypiwin32==223
pyrsistent==0.15.7
pytz==2019.3
pywin32==227
requests==2.22.0
rlp==1.2.0
semantic-version==2.8.4
six==1.14.0
sqlparse==0.3.0
toolz==0.10.0
typing-extensions==3.7.4.1
urllib3==1.25.8
varint==1.0.2
web3==5.4.0
websockets==8.1
zipp==2.0.1
